### 2025-07-16 608b41e
Dev of a detailed stat about selected pathway
