### 2025-07-18 33bb56a
Fix the code to display the page properly when not searching and link from hotbar
