### 2025-07-20 b4bec12
added plot of checkm QC for all bins of all samples in the bin page
