### 2025-10-12 #9188 @editorialbot
[REVIEW]: PANKEGG: Integrative Visualisation and Comparison of Metagenome-Assembled Genomes Annotation, Taxonomy, and Quality
**Submitting author:** @RVanDamme (<a href="http://orcid.org/0000-0002-7909-4868">Renaud Van Damme</a>)
**Repository:** https://github.com/RVanDamme/PANKEGG
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
