### 2025-10-12 d44d196
Use MAGs consistently in follow-up refs
