### 2025-10-12 bd6a898
Define KEGG at least once
