### 2025-10-12 6e558d1
Define SQL once
