### 2025-10-12 2d30d11
Prefer "macOS" over MacOS
