### 2025-10-28 264f8a4
Unified name as "PANKEGG"
Unified the name as "PANKEGG", while the usage of upper case or lower case is trivial matter, having it consistent in the paper matters most.
