+++ paper/paper.md
@@ -37,17 +37,17 @@ bibliography: paper.bib
 
 # Summary
 
-[Pankegg](https://github.com/RVanDamme/PANKEGG) is an interactive software suite for parsing and visualising metagenome-assembled genomes (MAGs) and exploring their metabolic capabilities. It integrates quality metrics, annotation, and taxonomic classification in one interactive central database.
-Pankegg enables researchers to explore, compare, and interpret their data through a modern browser-based interface, streamlining the analysis of large and complex metagenomic datasets. The software supports output from widely used tools such as CheckM2  [@Chklovski:2022], GTDB-TK [@Chaumeil:2020], Sourmash [@Brown:2016], and EggNOG [@Cantalapiedra:2021], making it a flexible solution for a wide range of microbiome and environmental genomics studies.
+[PANKEGG](https://github.com/RVanDamme/PANKEGG) is an interactive software suite for parsing and visualising metagenome-assembled genomes (MAGs) and exploring their metabolic capabilities. It integrates quality metrics, annotation, and taxonomic classification in one interactive central database.
+PANKEGG enables researchers to explore, compare, and interpret their data through a modern browser-based interface, streamlining the analysis of large and complex metagenomic datasets. The software supports output from widely used tools such as CheckM2  [@Chklovski:2022], GTDB-TK [@Chaumeil:2020], Sourmash [@Brown:2016], and EggNOG [@Cantalapiedra:2021], making it a flexible solution for a wide range of microbiome and environmental genomics studies.
 By interconnecting the different analysis results, people can investigate different interactions in the data more visually and conveniently. 
-Pankegg answers questions such as:
+PANKEGG answers questions such as:
 
 - How many of my bins pass the GTDB [@Parks:2022] quality threshold?
 - What is their taxonomic classification?
 - Which Kyoto Encyclopedia of Genes and Genomes (KEGG) orthologs are present?
 - Which proportion of their respective metabolic pathways is covered by the  KEGG orthologs identified? 
 
-Answering these questions typically requires inspecting multiple result files and cross-referencing the information. In contrast, Pankegg provides an all-in-one platform to explore, visualise, interpret, and save the results.
+Answering these questions typically requires inspecting multiple result files and cross-referencing the information. In contrast, PANKEGG provides an all-in-one platform to explore, visualise, interpret, and save the results.
 
 
 # Statement of need
@@ -55,15 +55,15 @@ Answering these questions typically requires inspecting multiple result files an
 The ever-growing progress of sequencing technologies has made it possible to recover thousands of draft and high-quality genomes directly from a plethora of environmental samples, accelerating our understanding of microbial diversity across ecosystems. Shotgun metagenomics with assembly-based approaches recovers MAGs, giving access to taxonomic and functional profiles of uncultured microorganisms.
 
 
-Pankegg integrates taxonomic analysis with KEGG pathway annotations, distinguishing itself from tools like MAGFlow/BIgMAG  [@YepesGarcia2024] and Anvi'o [@Eren2021]. While MAGFlow/BIgMAG and Anvi'o metagenomics provide comprehensive metagenomic workflows and their visualisation, Pankegg focuses on the visualisation and the functional interpretation of orthologs' variations across multiple samples and bins within the context of KEGG pathways [@Kanehisa:2023]. Our focused approach on the KEGG orthologs facilitates a more direct and efficient analysis of metabolic capabilities and variations in microbial communities, even with growing datasets. 
[…]
