### 2025-10-28 48b203a
Fixes of references from ISSUE #23
Fixed:
    "Anaconda, Inc..." I think you want to cite conda not Anaconda. You can get their preferred citation metadata from conda/CITATION.cff at main · conda/conda · GitHub.
    pip reference: I think a correct author for pip is "The pip developers", based on https://github.com/pypa/pip/blob/main/pyproject.toml
    python sqlite3 module: I think you want double curly braces for "Python Software Foundation"
    sqlite citation: You have "D. Richard Hipp" but I think it should be "Richard D. Hipp"
    pixi citation: Can you check what you have compared to th
[…]
