+++ paper/paper.bib
@@ -19,7 +19,7 @@
   number={1},
   pages={260},
   year={2022},
-  doi={10.1186/s13059-022-02809-9},
+  doi={10.1038/s41592-023-01940-w},
   issn={1474-760X},
   publisher={BioMed Central}
 }
@@ -75,7 +75,7 @@
 }
 
 @article{Eren2021,
-  author       = {A. Murat Eren and Evan Kiefl and Alon Shaiber and Iva Veseli and Samuel E. Miller and Matthew S. Schechter and Isaac Fink and Jessica N. Pan and Mahmoud Yousef and Emily C. Fogarty and Florian Trigodet and Andrea R. Watson and Ozcan C. Esen and Ryan M. Moore and Quentin Clayssen and Michael D. Lee and Veronika Kivenson and Elaina D. Graham and Bryan D. Merrill and Antti Karkman and Daniel Blankenberg and John M. Eppley and Andreas Sjodin and Jarrod J. Scott and Xabier Vazquez-Campos and Luke J. McKay and Elizabeth A. McDaniel and Sarah L. R. Stevens and Rika E. Anderson and Jessika Fuessel and Antonio Fernandez-Guerra and Lois Maignien and Tom O. Delmont and Amy D. Willis},
+  author       = { Murat A. Eren and Evan Kiefl and Alon Shaiber and Iva Veseli and Samuel E. Miller and Matthew S. Schechter and Isaac Fink and Jessica N. Pan and Mahmoud Yousef and Emily C. Fogarty and Florian Trigodet and Andrea R. Watson and Ozcan C. Esen and Ryan M. Moore and Quentin Clayssen and Michael D. Lee and Veronika Kivenson and Elaina D. Graham and Bryan D. Merrill and Antti Karkman and Daniel Blankenberg and John M. Eppley and Andreas Sjodin and Jarrod J. Scott and Xabier Vazquez-Campos and Luke J. McKay and Elizabeth A. McDaniel and Sarah L. R. Stevens and Rika E. Anderson and Jessika Fuessel and Antonio Fernandez-Guerra and Lois Maignien and Tom O. Delmont and Amy D. Willis},
   title        = {Community-led, integrated, reproducible multi-omics with anvi'o},
   journal      = {Nature Microbiology},
   year         = {2021},
@@ -87,7 +87,7 @@
 
 @manual{sqlite,
   title        = {SQLite},
-  author       = {D. Richard Hipp},
+  author       = {Hipp, {Richard D.} },
   organization = {SQLite Consortium},
   year         = {2000--2024},
   note         = {https://www.sqlite.org/}
@@ -95,13 +95,13 @@
 
 @manual{python-sqlite3,
   title        = {sqlite3 --- DB-API 2.0 interface for SQLite databases},
-  author       = {Python Software Foundation},
+  author       = {{Python Software Foundation}},
   year         = {2024},
   note         = {Python 3 Standard Library. \url{https://docs.python.org/3/library/sqlite3.html}}
 }
 
 @misc{flask,
-  author = {Armin Ronacher and contributors},
+  author = {Pallets},
   title = {Flask (Version 3.1.1)},
   year = {2024},
   howpublished = {\url{https://flask.palletsprojects.com/}},
@@ -110,7 +110,7 @@
 
 @article{reback2020pandas,
   title = {pandas-dev/pandas: Pandas (Version 2.2.3)},
-  author = {The pandas development team},
+  author = {{The pandas development team}},
   journal = {Zenodo},
   year = {2024},
   doi = {10.5281/zenodo.3509134},
@@ -131,7 +131,10 @@
 
 @article{scikit-learn,
   title = {Scikit-learn: Machine Learning in Python},
[…]
