### 2025-10-28 c793b2f
fixed typo in paper.bib preventing rendering
