+++ paper/paper.bib
@@ -181,7 +181,7 @@
 @misc{conda,
   title = {conda: A system-level, binary package and environment manager running on all major
   operating systems and platforms.},
-  author = {{conda contributors]},
+  author = {{conda contributors}},
   year = {2012},
   note = {Version 4.14.0},
   howpublished = {\url{https://docs.conda.io/projects/conda/}},
