### 2025-11-20 722c6c7
Update summary from @amoeba suggestion
