### 2025-11-24 c6da556
Fix secret key python-dotenv; updated installations; added scripts to generate input files
