+++ generate_input_gtdbtk_by_sample.sh
@@ -1,0 +1,8 @@
+echo "Sample name,Annotation_dir,classification_dir,Checkm2_dir" > samples.csv
+for sample in Dir/*; do
+   name=$(basename "$sample");
+   ann_path="$sample/bin_annotation/*.annotations.tsv";
+   class_path="$sample/gtdb_results/*.summary.tsv";
+   checkm2_path="$sample/checkm2_dir/quality_report.tsv";
+   echo "$name,$ann_path,$class_path,$checkm2_path" >> samples.csv;
+done
