### 2025-11-24 39832a3
Update readme from input of Reviewer Feedback. #25
Update readme install dependencies section to add "python-dotenv"
Edited the "generate input" files to call the new script files
Added how to generate the secret key to readme
