### 2025-11-24 3bdd82d
Update Manuscript from Issue 25 and Updated author list
