### 2025-11-30 #9188 @amoeba
@majensen my review is now an Accept based on the changes in https://github.com/RVanDamme/PANKEGG/tree/Review_correction and my checklist is complete.
