### 2025-12-17 #9188 @majensen
@RVanDamme thanks for the stimulus. Below are my copy edits of the paper itself, please review and make changes as you see fit. There is was an invalid DOI (detected by the bot) - I believe the correct one is there in my notes.

Can you please create an open archive of the codebase using Zenodo or similar. Ideally the archive authors and archive name will match those of the paper. Please provide the DOI you receive back here in the thread.

Copy review (line numbers in PDF proof):

l.63 (ref) and l. 179 - presume should read “Pandas Development Team”
l. 68 (ref) and l. 155 - should read “Python Software Foundation”
l. 69 (ref) and l. 131 - should read “Python Packaging Authority”

l. 85 - “many different information” -> “many different views on the data” (or similar)
l. 88-99 (numbered list): I would capitalize the first word following the colons.
l. 97: “The sample … Bins pages” -> “Sample vs. Sample, Bins vs. Bins pages: Allow users to…”, to be consistent with the other bullets
l.99: “This page visualises” -> “Visualizes”, to be consistent with the other bullets
l.101 - unformatted reference
l.104 - “and discovering” -> “and discovery of” (a list of nouns - generation, curation, discovery)
l. 118 (ref) - “Damme” -> “Van Damme” preferred? - also at l. 146
l. 129, l. 131, l. 172: URLs repeated in these refs.
l.142 Invalid DOI 10.1186/s13059-022-02809-9.
- It looks like the reference should be
	https://www.nature.com/articles/s41592-023-01940-w
	in Nature Methods, with DOI 10.1038/s41592-023-01940-w
l. 153: Capitalize Anvi’o ?
l. 159: Suggest adding the URL https://sqlite.org/consortium.html
