### 2025-12-17 bfdfc3d
Update paper.bib
attempt fixing the bib
