+++ paper/paper.bib
@@ -222,7 +222,7 @@ pages = {173--190}
 }
 
 @article{VanDamme2021,
-  author    = {Renaud {Van Damme} and Martin Hölzer and Adrian Viehweger and Bettina Müller and Erik {Bongcam-Rudloff} and Christian Brandt},
+  author    = {Renaud {Van~Damme} and Martin Hölzer and Adrian Viehweger and Bettina Müller and Erik {Bongcam-Rudloff} and Christian Brandt},
   title     = {Metagenomics workflow for hybrid assembly, differential coverage binning, metatranscriptomics and pathway analysis (MUFFIN)},
   journal   = {PLOS Computational Biology},
   volume    = {17},
