### 2025-12-17 bf2ee69
Update paper.bib
Update based on https://github.com/openjournals/joss-reviews/issues/9188#issuecomment-3663435457 comments
