+++ paper/paper.bib
@@ -91,11 +91,12 @@
   organization = {SQLite Consortium},
   year         = {2000--2024},
   note         = {https://www.sqlite.org/}
+  url          = {https://sqlite.org/consortium.html}
 }
 
 @manual{python-sqlite3,
   title        = {sqlite3 --- DB-API 2.0 interface for SQLite databases},
-  author       = {{Python Software Foundation}},
+  author       = {{Python~Software~Foundation}},
   year         = {2024},
   note         = {Python 3 Standard Library. \url{https://docs.python.org/3/library/sqlite3.html}}
 }
@@ -110,7 +111,7 @@
 
 @article{reback2020pandas,
   title = {pandas-dev/pandas: Pandas (Version 2.2.3)},
-  author = {{The pandas development team}},
+  author = {{The~pandas~development~team}},
   journal = {Zenodo},
   year = {2024},
   doi = {10.5281/zenodo.3509134},
@@ -119,7 +120,7 @@
 
 @article{harris2020array,
   title = {Array programming with {NumPy}},
-  author = {Harris, Charles R and Millman, K Jarrod and van der Walt, St\'efan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
+  author = {Harris, Charles R and Millman, K Jarrod and van~der~Walt, St\'efan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
   journal = {Nature},
   volume = {585},
   number = {7825},
@@ -171,20 +172,18 @@
 
 @misc{pip,
   title = {pip - The Python Package Installer},
-  author = {{The pip developers}},
+  author = {{The~pip~developers}},
   year = {2008},
   note = {Version 24.0},
-  howpublished = {\url{https://pip.pypa.io/}},
   url = {https://pip.pypa.io/}
 }
 
 @misc{conda,
   title = {conda: A system-level, binary package and environment manager running on all major
   operating systems and platforms.},
-  author = {{conda contributors}},
+  author = {{Conda~contributors}},
   year = {2012},
   note = {Version 4.14.0},
-  howpublished = {\url{https://docs.conda.io/projects/conda/}},
   url = {https://docs.conda.io/projects/conda/}
 }
 
@@ -192,7 +191,6 @@
   title = {Pixi: A modern package and environment manager for Python and Conda},
   author = {Ruben Arts and Bas Zalmstra and Wolf Vollprecht and Tim {de Jager} and Nichita Morcotilo and Julian Hofer},
   year = {2023},
-  howpublished = {\url{https://prefix.dev/docs/pixi/}},
   url = {https://prefix.dev/docs/pixi/}
 }
 
@@ -209,7 +207,7 @@
   publisher={Oxford University Press}
 }
 
-@article{Shaukat2016173190,
+@article{Shaukat:2016,
 author = {Shaukat, S. Shahid and Rao, Toqeer Ahmed and Khan, Moazzam A.},
 doi = {10.1515/eko-2016-0014},
 url = {https://doi.org/10.1515/eko-2016-0014},
