### 2025-12-17 2dec1e8
Update paper.md
Correcting paper.md based on comments in https://github.com/openjournals/joss-reviews/issues/9188#issuecomment-3663435457
