+++ paper/paper.md
@@ -100,18 +100,18 @@ This script ingests a comma-separated values file specifying the locations of Eg
 ## PANKEGG APP
 
 The app is a Flask-based web server that connects to the SQL database generated by `PANKEGG_make_db.py`,  providing a simple interface for interfacing, cross-referencing, and filtering the data.
-The interface offers many different information, and each page can be filtered by information from the other pages. A more detailed explanation is in our documentation's PANKEGG Web Page chapter.
+The interface offers many different views of the data, and each page can be filtered by information from the other pages. A more detailed explanation is in our documentation's PANKEGG Web Page chapter.
 Here is a concise summary:
 
-1. Bin page: outlines all bins/MAGs in the database. \autoref{fig:Figure1}
-2. Pathway page, also called Map page: lists the pathways in the database. Each pathway contains a “completion value”. This indicator is calculated by dividing the number of KEGG orthologs in the user's database by the total number of orthologs in the pathway.  \autoref{fig:Figure2}
-3. KEGG page: lists the KEGG orthologs in the data. The information for each ortholog is expandable, and the view then includes the corresponding EggNOG entries (bin ID, sample ID, GO terms, KEGG orthologs associated, EggNOG description).
-4. Taxonomy page: presents the taxonomic composition of the input database.
-5. The sample vs. Sample and Bins vs. Bins pages allow users to compare different samples or bins, respectively, with regard to pathway presence, completeness, and quality metrics.
-6. PCA page: This page visualizes a principal component analysis (PCA) based on functional or taxonomic profiles. Beware that PCA interpretation is only valid with enough data; we recommend at least 40 bins [Shaukat2016173190]. 
+1. Bin page: Outlines all bins/MAGs in the database. \autoref{fig:Figure1}
+2. Pathway page, also called Map page: Lists the pathways in the database. Each pathway contains a “completion value”. This indicator is calculated by dividing the number of KEGG orthologs in the user's database by the total number of orthologs in the pathway.  \autoref{fig:Figure2}
+3. KEGG page: Lists the KEGG orthologs in the data. The information for each ortholog is expandable, and the view then includes the corresponding EggNOG entries (bin ID, sample ID, GO terms, KEGG orthologs associated, EggNOG description).
+4. Taxonomy page: Presents the taxonomic composition of the input database.
+5. Sample vs. Sample, Bins vs. Bins pages: Allow users to compare different samples or bins, respectively, with regard to pathway presence, completeness, and quality metrics.
+6. PCA page: Visualises a principal component analysis (PCA) based on functional or taxonomic profiles. Beware that PCA interpretation is only valid with enough data; we recommend at least 40 bins [@Shaukat:2016]. 
 
 
-PANKEGG’s app is designed to make exploration intuitive. It features sortable and filterable tables, interactive plots, and external links to KEGG and other databases. These features collectively support users in hypothesis generation, genome curation, and discovering ecological and functional trends in complex datasets.
+PANKEGG’s app is designed to make exploration intuitive. It features sortable and filterable tables, interactive plots, and external links to KEGG and other databases. These features collectively support users in hypothesis generation, genome curation, and discovery of ecological and functional trends in complex datasets.
 
 # Figures
 
