### 2025-12-17 9fadf35
Update paper.bib
Fixed Anvi'o capitalisation.
