+++ paper/paper.bib
@@ -76,7 +76,7 @@
 
 @article{Eren2021,
   author       = { Murat A. Eren and Evan Kiefl and Alon Shaiber and Iva Veseli and Samuel E. Miller and Matthew S. Schechter and Isaac Fink and Jessica N. Pan and Mahmoud Yousef and Emily C. Fogarty and Florian Trigodet and Andrea R. Watson and Ozcan C. Esen and Ryan M. Moore and Quentin Clayssen and Michael D. Lee and Veronika Kivenson and Elaina D. Graham and Bryan D. Merrill and Antti Karkman and Daniel Blankenberg and John M. Eppley and Andreas Sjodin and Jarrod J. Scott and Xabier Vazquez-Campos and Luke J. McKay and Elizabeth A. McDaniel and Sarah L. R. Stevens and Rika E. Anderson and Jessika Fuessel and Antonio Fernandez-Guerra and Lois Maignien and Tom O. Delmont and Amy D. Willis},
-  title        = {Community-led, integrated, reproducible multi-omics with anvi'o},
+  title        = {Community-led, integrated, reproducible multi-omics with Anvi'o},
   journal      = {Nature Microbiology},
   year         = {2021},
   volume       = {6},
