### 2025-12-17 0ef3aee
Update paper.bib
