+++ paper/paper.bib
@@ -90,7 +90,7 @@
   author       = {Hipp, {Richard D.} },
   organization = {SQLite Consortium},
   year         = {2000--2024},
-  note         = {https://www.sqlite.org/}
+  note         = {https://www.sqlite.org/},
   url          = {https://sqlite.org/consortium.html}
 }
 
