### 2025-12-17 8dd2b8a
Fix typos paper.md
