+++ paper/paper.md
@@ -42,7 +42,7 @@ bibliography: paper.bib
 
 # Summary
 
-To study microorganisms present in an environment, many tools are needed. Each of these tools generates a large number of results files, providing different insights. [PANKEGG](https://github.com/RVanDamme/PANKEGG) is an all-in-one tool that centralises these results files and helps researachers navigate and visualize them for comparison, interpretation, and understanding of microbial communities in the environment. PANKEGG supports researchers by parsing and visualising metagenome-assembled genomes (MAGs) and exploring their metabolic capabilities. It integrates quality metrics, annotation, and taxonomic classification in one interactive central database.
+To study microorganisms present in an environment, many tools are needed. Each of these tools generates a large number of results files, providing different insights. [PANKEGG](https://github.com/RVanDamme/PANKEGG) is an all-in-one tool that centralises these results files and helps researchers navigate and visualise them for comparison, interpretation, and understanding of microbial communities in the environment. PANKEGG supports researchers by parsing and visualising metagenome-assembled genomes (MAGs) and exploring their metabolic capabilities. It integrates quality metrics, annotation, and taxonomic classification in one interactive central database.
 
 PANKEGG enables researchers to explore, compare, and interpret their data through a modern browser-based interface, streamlining the analysis of large and complex metagenomic datasets. The software supports output from widely used tools such as CheckM2  [@Chklovski:2022], GTDB-TK [@Chaumeil:2020], Sourmash [@Brown:2016], and EggNOG [@Cantalapiedra:2021], making it a flexible solution for a wide range of microbiome and environmental genomics studies.
 By interconnecting the different analysis results, researchers can investigate different interactions in the data more visually and conveniently.
