### 2025-12-18 9340a05
Create CITATION.cff
