### 2025-12-22 3ec51c0
Update paper.md
