### 2025-12-22 150f277
Update paper.md
Fixing "Van Damme" name in citation on footnote
