### 2026-01-19 bc6fa4d
Update README.md
Added support chapter
