### 2026-01-21 #9188 @arfon
> @openjournals/dev @arfon is still resolved now? Should I reaccept or can we close this? Thanks!

OK, this should all be fixed now.
